requirements.txt
python-dotenv
langchain 
langchain_community
langchain_openai
streamlit 